+++ src/components/RainbowKit/RainbowKit.tsx
@@ -1,0 +1,42 @@
+import React, { ReactNode } from 'react'
+import '@rainbow-me/rainbowkit/styles.css';
+import {
+  getDefaultWallets,
+  RainbowKitProvider,
+} from '@rainbow-me/rainbowkit';
+import { goerli, configureChains, createClient, WagmiConfig } from 'wagmi';
+import { publicProvider } from 'wagmi/providers/public';
+
+const { chains, provider } = configureChains(
+  [
+    goerli,
+  ],
+  [publicProvider()]
+);
+
+const { connectors } = getDefaultWallets({
+  appName: 'Extent App',
+  chains,
+});
+
+const wagmiClient = createClient({
+  autoConnect: true,
+  connectors,
+  provider,
+});
+
+interface Props {
+    children?: ReactNode
+}
+
+const Rainbow: React.FC<Props> = ({ children }) => {
+  return (
+    <WagmiConfig client={wagmiClient}>
+      <RainbowKitProvider chains={chains}>
+        {children}
+      </RainbowKitProvider>
+    </WagmiConfig>
+  );
+};
+
+export default Rainbow;
